Huge code cleanup part 2.
after 2-3 hours finally finished :pray:

+++ src/main/java/me/earth/phobos/features/modules/misc/AutoReconnect.java
@@ -82,7 +82,7 @@ class AutoReconnect
 
         public
         void updateScreen ( ) {
-            if ( this.timer.passedS ( AutoReconnect.this.delay.getValue ( ).intValue ( ) ) ) {
+            if ( this.timer.passedS ( AutoReconnect.this.delay.getValue ( ) ) ) {
                 this.mc.displayGuiScreen ( new GuiConnecting ( this.parentScreen , this.mc , serverData == null ? this.mc.currentServerData : serverData ) );
             }
         }
